--- packages/vite/src/types/lightningcss.d.ts
@@ -1,16 +1,0 @@
-import type {
-  BundleAsyncOptions,
-  CustomAtRules,
-  TransformAttributeOptions,
-} from 'lightningcss'
-
-// remove options overriden by vite
-export type LightningCSSOptions = Omit<
-  BundleAsyncOptions<CustomAtRules> & TransformAttributeOptions,
-  | 'filename'
-  | 'code'
-  | 'resolver'
-  | 'minify'
-  | 'sourceMap'
-  | 'analyzeDependencies'
->
